improvement : create pelatihanku pages
