+++ src/pages/pelatihanku/modul/[slug]/index.tsx
@@ -1,4 +1,4 @@
-import RegistrationVideoModule from "@/modules/pelatihanku/registration-video/MyTrainingVideoModule";
+import RegistrationVideoModule from "@/modules/mytraining-video-module/MyTrainingVideoModule";
 import { NextPage } from "next";
 import React, { ReactElement, Suspense } from "react";
